package/mono: bump to version 5.0.1.1 and monolite to version 1050001000
This patch bumps to mono 5.0.1.1 and updates the unit tests patch
accordingly.

The latest version of mono carries a bit of changes in the monolite
package: the version string changes and from now on, monolite should
be installed in a subdirectory with the exact version string as a name.
Therefore, mono and monolite are bumped in a single commit.

Patch 0004, although applied in upstream master, is not yet in this
(stable) release.

Signed-off-by: Angelo Compagnucci <angelo.compagnucci@gmail.com>
[Arnout: squashed the patches, removed /usr part from HOST_DIR]
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>