boarde: icorem6: clarify the common vs. board-specific things in readme.txt
The parts of the readme.txt that specify the board-specific aspects
(DTB files, KMS config json file) were a bit confusing. For the DTB
files this is solved by putting the board it applies to after the DTB
file name. For the json files, this is solved by adding an extra line,
so the CinematicExperience-demo is clearly separated from the setting
of QT_QPA_EGLFS_KMS_CONFIG.

Also remove the 'for all boards' line in the beginning, it doesn't add
any value.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>