socat: bump down to version 1.7.3.2
Upstream socat2 branch seems to be dead. Last commit is from January
2016 . Over the last few years socat2 received only fixes cherry picked
from the master 1.x branch. Most major general purpose distros only
package socat 1.x.

Drop the threads dependency; not needed for 1.x.

Mention the OpenSSL exception in the license. Add hashes for license
files.

Correct the autoconf comment. The tarball configure script is recent
enough. But since we patch configure.in and Makefile.in we still need to
autoconf.

All patches we carry for socat2 are also needed for socat 1.x.

Signed-off-by: Baruch Siach <baruch@tkos.co.il>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>