webkitgtk: security bimp to version 2.18.4
This is a maintenance release of the current stable WebKitGTK+ version,
which contains fixes for CVE-2017-13866, CVE-2017-13870, CVE-2017-7156, and
CVE-2017-13856.  Additionally, this release brings improvements in the
WebDriver spec-compliance, plugs several memory leaks in its GStreamer based
multimedia backend, and fixes a bug when handling cookie removal.

Release notes can be found in the announcement:

  https://webkitgtk.org/2017/12/19/webkitgtk2.18.4-released.html

More details about the security fixes are provided in the following
WebKitGTK+ Security Advisory report:

  https://webkitgtk.org/security/WSA-2017-0010.html

Last but not least, this new release includes the fix for honoring the
CMAKE_BUILD_TYPE value from CMake toolchain files and the corresponding
patch is removed.

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>